ID validation, improved code, added more validations.

+++ app/Http/Controllers/V1/SecretController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\V1;
 use App\Http\Controllers\Controller;
 use App\Services\SecretService;
 use Carbon\Carbon;
+use http\Env\Response;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
@@ -13,6 +14,8 @@ class SecretController extends Controller
 {
 
     private SecretService $secretService;
+
+    private int $default_expire_message_in_hours = 7;
 
     public function __construct(SecretService $secretService)
     {
@@ -26,12 +29,19 @@ class SecretController extends Controller
     public function add(Request $request): JsonResponse
     {
 
+        $id = $request->input("id");
+        $message = $request->input("message");
+
+        if(strlen($id) !== 36 || $id === null) {
+            return response()->json([]);
+        }
+
         // in hours.
         $expires_at = (int) $request->input('expires_at');
 
         // if none is set, expire in 7 days.
-        if($expires_at == 0) {
-            $expires_at = 24 * 7;
+        if($expires_at <= 0) {
+            $expires_at = 24 * $this->default_expire_message_in_hours;
         }
 
         $expires_at = Carbon::createFromFormat('Y-m-d H:i:s', date("Y-m-d H:i:s"))->addHours($expires_at);
@@ -43,10 +53,10 @@ class SecretController extends Controller
         }
 
         $data = [
-            'id' => $request->input('id'),
-            'message' => $request->input('message'),
-            'expires_at' => $expires_at,
-            'password' => $password
+            'id'            => $id,
+            'message'       => $message,
+            'expires_at'    => $expires_at,
+            'password'      => $password
         ];
 
         $secret = $this->secretService->add($data)->object();
@@ -61,11 +71,13 @@ class SecretController extends Controller
     public function view(Request $request): JsonResponse
     {
 
-        if( $request->input('id') === null) {
+        $id = $request->input('id');
+
+        if($id === null) {
             return response()->json(['response_code' => 400]);
         }
 
-        $id = hash("sha512", $request->input('id'));
+        $id = hash("sha512", $id);
 
         $find = $this->secretService->view($id)->object();
 
@@ -86,8 +98,15 @@ class SecretController extends Controller
      */
     public function delete(Request $request): JsonResponse
     {
+
+        $id = $request->input('id');
+
+        if($id === null) {
+            return response()->json(['response_code' => 400]);
+        }
+
         $delete = $this->secretService->delete(
-            hash("sha512", $request->input('id'))
+            hash("sha512", $id)
         )->object();
 
         return response()->json($delete);
